+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - **_Breaking Change_**: rewrite `ASTNode`, add `NTNode` and `TNode`.
     - Rewrite selectors and traversers.
   - Feat: add `ExtractSerializableParserData`. [#41](https://github.com/DiscreteTom/retsac/issues/41)
+  - Feat: add `NTNode.global` to share data across all NT nodes.
 - Perf: prevent unnecessary calculation.
 - Reduce package size. [#31](https://github.com/DiscreteTom/retsac/issues/31)
   - Optimize the format of serialized parser data, remove unused strings.